ref(replay): Move ReplayPlayerTimestampEmitter into its own file
